Categorize monitoring tools
The current list of opensource and hosted monitoring tools has a great selection, but it can be a bit confusing knowing the focus area of each tool. I've categorized the existing tools (did not change the actual tools listed) to make it a bit easier to understand the speciality area of each tool.